requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
beautifulsoup4==4.11.1
boto3==1.26.26
botocore==1.29.26
certifi==2022.9.24
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.3
django-cors-headers==3.13.0
django-restframework==0.0.1
djangorestframework==3.14.0
durationpy==0.5
favicon==0.7.0
filelock==3.8.2
gunicorn==20.1.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
nodejs==0.1.1
optional-django==0.1.0
Pillow==9.3.0
placebo==0.9.0
platformdirs==2.6.0
psycopg==3.1.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pypi==2.1
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.2.0
tzdata==2022.6
urllib3==1.26.12
virtualenv==20.17.1
Werkzeug==2.2.2
whitenoise==6.2.0
zappa==0.56.1
